+++ packages/icons/src/index.ts
@@ -385,7 +385,8 @@ export { Speed } from './speed';
 export { Mute } from './mute';
 export { VariantPlaceholder } from './variant-placeholder';
 export { NoLink2 } from './no-link-2';
-export { NoLabels2 } from './no-labels-2';
+export { NoProducts as NoLabels2 } from './no-products'; // to be removed to avoid duplicates
+export { NoProducts } from './no-products';
 export { NoKeys2 } from './no-keys-2';
 export { Any } from './any';
 export { Dark } from './dark';