+++ packages/ui/src/core/date-picker/date-picker.component.tsx
@@ -22,6 +22,7 @@ const Datepicker = intrinsicComponent<DatePickerProps, HTMLDivElement>(
       value = '',
       onChange,
       autoSelectToday,
+      enableAutoSelect,
       maxDate = '',
       minDate = '',
       position,
@@ -41,7 +42,6 @@ const Datepicker = intrinsicComponent<DatePickerProps, HTMLDivElement>(
     const [inputValue, setInputValue] = useState(value);
     const [showPlaceholder, setShowPlaceholder] = useState(true);
     const [isHovering, setIsHovering] = useState(false);
-
     const datePickerRef = useRef(null);
 
     const maxDateTimestamp = new Date(maxDate).getTime();
@@ -126,6 +126,7 @@ const Datepicker = intrinsicComponent<DatePickerProps, HTMLDivElement>(
           maxDate={maxDate}
           minDate={minDate}
           autoSelectToday={autoSelectToday}
+          enableAutoSelect={enableAutoSelect}
           anchorEl={datePickerRef.current}
           position={position || 'bottom-start'}
           popperOptions={popperOptions}
@@ -138,6 +139,7 @@ const Datepicker = intrinsicComponent<DatePickerProps, HTMLDivElement>(
 
 Datepicker.defaultProps = {
   calendarStyles: {},
+  enableAutoSelect: false,
 };
 
 export const propTypes = {
@@ -151,6 +153,7 @@ export const propTypes = {
   hint: PT.string,
   popperOptions: popperPropTypes.popperOptions,
   autoSelectToday: PT.bool,
+  enableAutoSelect: PT.bool,
   fullWidth: PT.bool,
   InputProps: PT.exact(inputPropTypes) as Validator<InputProps>,
   inputGroupProps: PT.exact(inputPropTypes) as Validator<InputGroupProps>,