+++ packages/ui/src/core/modal-title/modal-title.component.tsx
@@ -8,11 +8,13 @@ import Styled from './modal-title.styles';
 import { Variant } from './types';
 
 const ModalTitle = intrinsicComponent<ModalTitleProps, HTMLDivElement>(
-  ({ secondary, onClose, ...rest }, ref): JSX.Element => (
+  ({ secondary, onClose, primaryLabelStyles, ...rest }, ref): JSX.Element => (
     <Styled.ModalTitle {...rest} ref={ref}>
       {rest.icon && <Styled.Icon iconShadow={Boolean(rest.iconShadow)}>{rest.icon}</Styled.Icon>}
 
-      <Styled.LabelPrimary {...rest}>{rest.primary}</Styled.LabelPrimary>
+      <Styled.LabelPrimary variant={rest.variant} style={primaryLabelStyles}>
+        {rest.primary}
+      </Styled.LabelPrimary>
 
       {rest.variant === Variant.WithIcon && secondary && <Styled.LabelSecondary>{secondary}</Styled.LabelSecondary>}
 
@@ -34,6 +36,7 @@ ModalTitle.propTypes = {
   icon: PT.node,
   iconShadow: PT.bool,
   onClose: PT.func,
+  primaryLabelStyles: PT.object,
   variant: PT.oneOf(objectValues(Variant)),
 };