+++ packages/ui/stories/core/autocomplete.stories.tsx
@@ -1,7 +1,12 @@
 import React, { useState } from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
+import NoResult from '@scaleflex/icons/no-result';
+import { AutocompleteValueType } from '@scaleflex/ui/core/autocomplete/autocomplete.props';
+
 import Autocomplete from '../../src/core/autocomplete';
 import { InputSize } from '../../src/utils/types';
+import AiIcon from './icons/ai-icon';
+import Styled from './autocomplete.story.styled';
 
 const meta: Meta<typeof Autocomplete> = {
   title: 'Inputs/Autocomplete',
@@ -40,32 +45,21 @@ const defaultArgs = {
   multiple: false,
   fullWidth: false,
 };
-
-const options = [
-  { id: 1, label: 'sfx1' },
-  { id: 2, label: 'sfx2' },
-  { id: 3, label: 'sfx3' },
-  { id: 4, label: 'sfx4' },
-  { id: 5, label: 'sfx5' },
-  { id: 6, label: 'sfx6' },
-  { id: 7, label: 'sfx7' },
-];
-
-const getOptionDisabled = (_: any, index: number): boolean => index % 2 === 0;
-
 const BasicTemplate = ({ ...args }): JSX.Element => {
-  const [value, setValue] = useState(args.multiple ? [] : '');
+  const [value, setValue] = useState<AutocompleteValueType>(args.multiple ? [] : '');
 
   return (
-    <Autocomplete
-      {...args}
-      value={value}
-      options={args.options}
-      getOptionDisabled={getOptionDisabled}
-      getOptionValue={(option: any) => option}
-      getOptionLabel={(option: any) => option}
-      onChange={(_: any, val: any) => setValue(val)}
-    />
+    <Styled.Container fullWidth={Boolean(args.fullWidth)}>
+      <Autocomplete
+        {...args}
+        value={value}
+        options={args.options}
+        getOptionValue={(option: any) => option}
+        getOptionLabel={(option: any) => option}
+        getOptionDisabled={(option: any) => option === defaultArgs.options[3]}
+        onChange={setValue}
+      />
+    </Styled.Container>
   );
 };
 
@@ -77,13 +71,53 @@ const AutocompleteObjectsTemplate = ({ ...args }): JSX.Element => {
       {...args}
       value={value}
       options={args.options}
-      onChange={(_, val: any) => setValue(val)}
-      getOptionValue={(option: any) => option?.id}
-      getOptionLabel={(option: any) => option?.label}
-      getOptionDisabled={getOptionDisabled}
+      onChange={setValue}
+      getOptionValue={(option: any) => option?.uuid}
+      getOptionLabel={(option: any) => option?.name}
+      getOptionDisabled={(option: any) => option.uuid === defaultArgs.options[3]}
     />
   );
 };
+
+const WithIconsTemplate = ({ ...args }): JSX.Element => {
+  const [value, setValue] = useState<AutocompleteValueType>(args.multiple ? [] : '');
+
+  return (
+    <Styled.Container fullWidth={Boolean(args.fullWidth)}>
+      <Autocomplete
+        {...args}
+        value={value}
+        options={args.options}
+        onChange={setValue}
+        getOptionValue={(option: any) => option?.uuid}
+        getOptionLabel={(option: any) => option?.name}
+        getOptionDisabled={(option: any) => option?.disabled}
+        renderSearchEmptyMenuItem={() => (
+          <Styled.NoSearchResultsContainer>
+            <NoResult width={150} size={60} />
+            <div>Try another search.</div>
+          </Styled.NoSearchResultsContainer>
+        )}
+        renderLabelIconEnd={({ option }: any) => option?.endIcon}
+        renderOptionLabel={(option: any) => (
+          <Styled.OptionLabel>
+            {option.name}
+            <Styled.OptionLabelIconWrapper>{option?.endIcon}</Styled.OptionLabelIconWrapper>
+          </Styled.OptionLabel>
+        )}
+      />
+    </Styled.Container>
+  );
+};
+
+const optionsWithIconsAndFavorite = [
+  { uuid: 'en', name: 'English', endIcon: <AiIcon />, favorite: true },
+  { uuid: 'fr', name: 'French', endIcon: <AiIcon />, favorite: true },
+  { uuid: 'de', name: 'German', favorite: true },
+  { uuid: 'it', name: 'Italian' },
+  { uuid: 'es', name: 'Spanish', disabled: true },
+  { uuid: 'ua', name: 'Ukrainian', endIcon: <AiIcon /> },
+];
 
 export const Primary: Story = {
   args: { ...defaultArgs },
@@ -96,4 +130,14 @@ export const AutocompleteObjects: Story = {
     options,
   },
   render: (args) => <AutocompleteObjectsTemplate {...args} />,
+};
+
+export const WithIcons: Story = {
+  args: { ...defaultArgs, label: 'Languages', placeholder: 'Select language', options: optionsWithIconsAndFavorite },
+  render: (args) => <WithIconsTemplate {...args} />,
+};
+
+export const WithGroups: Story = {
+  args: { ...WithIcons.args, groupBy: (option: any) => option?.favorite },
+  render: (args) => <WithIconsTemplate {...args} />,
 };